Remove contexts
